refactor: 컨버터에서 웹소켓 전용 로직은 별도 클래스로 분리
